Use Int instead of Long for Ids.

+++ plugins/woocommerce/src/main/kotlin/org/wordpress/android/fluxc/persistence/dao/SSRDao.kt
@@ -13,5 +13,5 @@ abstract class SSRDao {
     abstract suspend fun insertSSR(ssrEntity: SSREntity): Long
 
     @Query("SELECT * FROM SSREntity WHERE localSiteId = :localSiteId LIMIT 1")
-    abstract suspend fun getSSRbySite(localSiteId: Long): WCSSRModel
+    abstract suspend fun getSSRbySite(localSiteId: Int): WCSSRModel
 }
